requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
celery==5.2.7
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chatterbot-corpus==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constants==0.6.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.18.0
Flask==3.0.0
flask-cors==5.0.1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.30.2
idna==3.7
ipykernel==6.29.4
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
Js2Py==0.74
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
kombu==5.5.3
Levenshtein==0.27.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.4.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
openai==1.78.0
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
phonenumbers==9.0.4
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
pipwin==0.5.2
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.0.0
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.27.1
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
RapidFuzz==3.13.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
selenium==4.23.1
Send2Trash==1.8.3
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.5
SpeechRecognition==3.14.1
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.26.2
trio-websocket==0.11.1
twilio==8.3.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
utils==1.0.2
vine==5.1.0
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.20.0
